+++ src/pages/home/Home.style.tsx
@@ -1,0 +1,29 @@
+import styled from "styled-components";
+
+export const MainContainer = styled.div`
+  padding-top: 50px;
+  height: 100vh;
+  padding-bottom: 10rem;
+  background-color: #fafafa;
+  font-family: "SUIT", sans-serif;
+  overflow-x: hidden;
+
+  display: flex;
+  align-items: center;
+  flex-direction: column;
+`;
+
+export const FabContainer = styled.div`
+  position: fixed;
+
+  display: flex;
+  width: 100vw;
+  justify-content: center;
+  height: 5rem;
+  bottom: 0rem;
+
+  display: flex;
+  justify-content: center;
+
+  z-index: 5;
+`;
